% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_in}
\alias{is_in}
\title{Check if an instance is a member of a list or vector}
\usage{
is_in(instance, vector_or_list)
}
\arguments{
\item{instance}{an R6 instance}

\item{vector_or_list}{a vector or list (possibly empty)}
}
\value{
TRUE of FALSE
}
\description{
Check if an instance is a member of a list or vector
}
\author{
fnaufel
}
